fix(cli_runners): use MagicMock for sync proc.kill; widen timeout_s to float
proc.kill() is a synchronous method — AsyncMock returned an un-awaited
coroutine that generated RuntimeWarning. Replaced with MagicMock.
Widened timeout_s annotation from int to float so passing 0.05 in tests
matches the type signature (asyncio.wait_for already accepts float|None).

+++ agents-sdk/lib/cli_runners.py
@@ -86,7 +86,7 @@ CODEX_BINARY = "/opt/homebrew/bin/codex"
 CODEX_DEFAULT_TIMEOUT_S = 120
 
 
-async def run_codex(prompt: str, timeout_s: int = CODEX_DEFAULT_TIMEOUT_S) -> CLIResponse:
+async def run_codex(prompt: str, timeout_s: float = CODEX_DEFAULT_TIMEOUT_S) -> CLIResponse:
     """Invoke `codex exec` with read-only sandbox and skip-git-repo-check.
 
     Runs from `Path.home()` (trusted per ~/.codex/config.toml `[projects]`).
